# apps/orders-api/openapi.yaml
openapi: 3.0.3
info:
  title: Orders API
  description: |
    API REST para gestión de órdenes de compra, productos y stock.
    Parte del sistema de orquestación de pedidos B2B.
  version: 1.0.0

tags:
  - name: Health
    description: Endpoints de monitoreo
  - name: Orders
    description: Gestión de órdenes
  - name: Products
    description: Gestión de productos

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Verifica el estado del servicio
      responses:
        '200':
          description: Servicio operativo
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: OK
                  message:
                    type: string
                    example: Orders API is running
                  timestamp:
                    type: string
                    format: date-time
                    example: '2026-02-09T01:55:19.854Z'

  /admin/orders/create-order:
    post:
      tags:
        - Orders
      summary: Crear nueva orden
      description: |
        Crea una nueva orden validando el cliente, verificando stock disponible,
        calculando totales y descontando el stock de productos.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            examples:
              simple:
                summary: Orden simple con un producto
                value:
                  customer_id: 1
                  items:
                    - product_id: 2
                      qty: 3
              multiple:
                summary: Orden con múltiples productos
                value:
                  customer_id: 1
                  items:
                    - product_id: 1
                      qty: 2
                    - product_id: 3
                      qty: 1
      responses:
        '201':
          description: Orden creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Datos inválidos o stock insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Cliente o producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /admin/orders/find-all:
    get:
      tags:
        - Orders
      summary: Listar órdenes
      description: Obtiene lista paginada de órdenes con filtros opcionales
      parameters:
        - name: status
          in: query
          schema:
            type: integer
            enum: [1, 2, 3]
          description: 'Estado de la orden (1: CREATED, 2: CONFIRMED, 3: CANCELED)'
        - name: from
          in: query
          schema:
            type: string
            format: date
          description: Fecha desde (YYYY-MM-DD)
        - name: to
          in: query
          schema:
            type: string
            format: date
          description: Fecha hasta (YYYY-MM-DD)
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
          description: Cantidad de resultados
        - name: cursor
          in: query
          schema:
            type: integer
          description: ID desde donde continuar la paginación
      responses:
        '200':
          description: Lista de órdenes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'

  /admin/orders/find-one/{id}:
    get:
      tags:
        - Orders
      summary: Obtener orden por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalle de la orden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Orden no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /admin/orders/{id}/confirm:
    post:
      tags:
        - Orders
      summary: Confirmar orden
      description: |
        Confirma una orden cambiando su estado de CREATED a CONFIRMED.
        Implementa idempotencia mediante el header X-Idempotency-Key.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: X-Idempotency-Key
          in: header
          required: true
          schema:
            type: string
          description: Clave única para garantizar idempotencia
          example: abc-123-def-456
      responses:
        '200':
          description: Orden confirmada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: La orden no puede ser confirmada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /admin/orders/{id}/cancel:
    post:
      tags:
        - Orders
      summary: Cancelar orden
      description: |
        Cancela una orden y restaura el stock de productos.
        Reglas: CREATED se puede cancelar siempre, CONFIRMED solo dentro de 10 minutos.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Orden cancelada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: La orden no puede ser cancelada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CreateOrderRequest:
      type: object
      required:
        - customer_id
        - items
      properties:
        customer_id:
          type: integer
          minimum: 1
          description: ID del cliente
          example: 1
        items:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - product_id
              - qty
            properties:
              product_id:
                type: integer
                minimum: 1
                description: ID del producto
                example: 2
              qty:
                type: integer
                minimum: 1
                description: Cantidad del producto
                example: 3

    OrderResponse:
      type: object
      properties:
        errorCode:
          type: integer
          example: 0
        message:
          type: string
          example: Orden creada exitosamente
        data:
          $ref: '#/components/schemas/Order'

    Order:
      type: object
      properties:
        id:
          type: integer
          example: 101
        customerId:
          type: integer
          example: 1
        status:
          type: integer
          enum: [1, 2, 3]
          description: '1: CREATED, 2: CONFIRMED, 3: CANCELED'
          example: 2
        totalCents:
          type: integer
          description: Total en centavos
          example: 459900
        cantProducts:
          type: integer
          description: Cantidad total de productos
          example: 3
        nroOrder:
          type: string
          description: Número de orden
          example: ORD-7437692854
        createdAt:
          type: string
          format: date-time
          example: '2026-02-09T01:57:41.000Z'
        orderProducts:
          type: array
          items:
            $ref: '#/components/schemas/OrderProduct'

    OrderProduct:
      type: object
      properties:
        id:
          type: integer
          example: 1
        orderId:
          type: integer
          example: 101
        productId:
          type: integer
          example: 2
        qty:
          type: integer
          example: 3
        unitPriceCents:
          type: integer
          example: 129900
        subtotalCents:
          type: integer
          example: 389700
        product:
          $ref: '#/components/schemas/Product'

    Product:
      type: object
      properties:
        id:
          type: integer
          example: 2
        sku:
          type: string
          example: PROD-001
        name:
          type: string
          example: Producto de ejemplo
        priceCents:
          type: integer
          example: 129900
        stock:
          type: integer
          example: 47
        createdAt:
          type: string
          format: date-time

    OrderListResponse:
      type: object
      properties:
        errorCode:
          type: integer
          example: 0
        message:
          type: string
          example: Órdenes encontradas
        data:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        meta:
          type: object
          properties:
            limit:
              type: integer
              example: 20
            cursor:
              type: integer
              nullable: true
              example: null
            nextCursor:
              type: integer
              nullable: true
              example: 115
            hasNext:
              type: boolean
              example: true

    ErrorResponse:
      type: object
      properties:
        errorCode:
          type: integer
          example: 500
        message:
          type: string
          example: Error al crear la orden
        data:
          nullable: true